Rename auth database driver namespace
